requirements.txt
allure-pytest>=2.10.0
selenium>=4.15.0
pytest>=7.4.0
python-dotenv>=1.0.0
Faker>=20.0.0
pytest-rerunfailures>=12.0
pytest-xdist>=3.5.0